requirements.txt
langchain
langchain-openai
openai
python-dotenv
tiktoken
pypdf